requirements.txt
altair==4.2.2
ansicon==1.89.0
anyio==3.6.2
attrs==23.1.0
Babel==2.12.1
blessed==1.20.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.0.4
codefind==0.1.3
colorama==0.4.6
decorator==5.1.1
docxcompose==1.4.0
docxtpl==0.16.0
entrypoints==0.4
fastapi==0.94.1
Flask==2.2.3
Flask-Cors==3.0.10
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
jsonschema==4.17.3
jurigged==0.5.5
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mysql-connector-python==8.0.33
numpy==1.21.5
ovld==0.3.2
packaging==23.1
pandas==1.4.2
Pillow==9.5.0
plotly==5.7.0
protobuf==3.19.6
pyarrow==11.0.0
pydantic==1.10.6
pydeck==0.8.1b0
pyFirmata==1.1.0
Pympler==1.0.1
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
semver==3.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
streamlit==1.9.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.3
zipp==3.15.0
